refactor and pass in past data
